requirements.txt
youtube-transcript-api>=0.6.0
google-generativeai>=0.8.0
discord.py>=2.3.0
streamlit>=1.30.0
python-dotenv>=1.0.0
Pillow>=10.0.0
